refactor: lift antd layout components from components to pages
Lift `antd` Layout components (Layout and Sider) to parent
`Courses.tsx`

This way, components related to courses only need to focus on the
body of the page and are not concerned with the markup around them
